+++ tests/test_set/mod.rs
@@ -1,0 +1,11 @@
+
+#[cfg(test)]
+pub mod gen_tests;
+
+#[cfg(test)]
+pub mod expr_tests;
+
+#[cfg(test)]
+pub mod stmt_tests;
+
+
